[Pri] Add k-times OP
Use midicsv2ass to generate k-timed ASS lines.

Prepare the romaji and english files to be suitable for midicsv2ass input.

Also include the source MuseScore file that generated the MIDI file.

The English has also been split into syllables, although this output has
not been included in the OP .ass file. The k-timed ASS files for the
English lyrics are in a separate file.